+++ public/components/create_report/create_report.js
@@ -39,97 +39,11 @@ import {
   } from '@elastic/eui';
 import { htmlIdGenerator } from '@elastic/eui/lib/services';
 import moment from 'moment';
-import { thisExpression } from '@babel/types';
 import { ReportSettings } from './report_settings/report_settings';
 import { ReportDelivery } from './delivery/delivery';
 import { ReportSchedule } from './schedule/schedule';
 
 const idPrefix = htmlIdGenerator()();
-
-const DeliveryChannelRadio = () => {
-  const radios = [
-    {
-      id: `${idPrefix}3`,
-      label: 'None (report will be available in Reports List page)',
-    },
-    {
-      id: `${idPrefix}4`,
-      label: 'Email',
-    },
-    {
-      id: `${idPrefix}5`,
-      label: 'Chime',
-    },
-    {
-      id: `${idPrefix}6`,
-      label: 'Other (webhook)'
-    }
-  ];
-  
-  const [radioIdSelected, setRadioIdSelected] = useState(`${idPrefix}3`);
-
-  const onChangeDeliveryChannel = optionId => {
-    setRadioIdSelected(optionId);
-  };
-  
-  return (
-      <EuiRadioGroup
-        options={radios}
-        idSelected={radioIdSelected}
-        onChange={id => onChangeDeliveryChannel(id)}
-        name="delivery channel radio group"
-        legend={{
-          children: <span>This is a legend for a radio group</span>,
-        }}
-      />
-  );
-};
-
-const DeliveryRecipientsBox = () => {
-  const options = [];
-  const [selectedOptions, setSelected] = useState([]);
-
-  const onChangeDeliveryRecipients = selectedOptions => {
-    setSelected(selectedOptions);
-  };
-
-  const onCreateDeliveryRecipientOption = (searchValue, flattenedOptions = []) => {
-    const normalizedSearchValue = searchValue.trim().toLowerCase();
-  
-    if (!normalizedSearchValue) {
-      return;
-    }
-  
-    const newOption = {
-      label: searchValue,
-    };
-  
-    // Create the option if it doesn't exist.
-    if (
-      flattenedOptions.findIndex(
-      option => option.label.trim().toLowerCase() === normalizedSearchValue
-      ) === -1
-    ) {
-      options.push(newOption);
-    }
-  
-    // Select the option.
-    setSelected([...selectedOptions, newOption]);
-  };
-
-  return (
-    <EuiComboBox
-      placeholder="Select or create options"
-      options={options}
-      selectedOptions={selectedOptions}
-      onChange={onChangeDeliveryRecipients}
-      onCreateOption={onCreateDeliveryRecipientOption}
-      isClearable={true}
-      data-test-subj="demoComboBox"
-    />
-  );
-};
-
 
 export class CreateReport extends React.Component {
   constructor(props) {
@@ -149,12 +63,6 @@ export class CreateReport extends React.Component {
       scheduleRecurringWeeklyDayOfWeek: 'Monday',
       scheduleRecurringStartDate: moment(),
     };
-  }
-
-  onChangeReportSettingsRadio = (e) => {
-      // this.setState({
-      //     reportSettingsRadioIdSelected: e
-      // });
   }
   
   onChangeReportSettingsDashboard = (e) => {
@@ -270,13 +178,14 @@ export class CreateReport extends React.Component {
         options={radios}
         idSelected={radioIdSelected}
         onChange={id => onChangeSettingsRadio(id)}
-        name="schedule radio group"
+        name="scheduleRadioGroup"
         legend={{
           children: <span>This is a legend for a radio group</span>,
         }}
       />
     );
   };
+  
 
   timezone_options = [
       { value: -4, text: 'EDT -04:00' },
@@ -472,8 +381,6 @@ export class CreateReport extends React.Component {
           />
           <EuiSpacer/>
           <ReportDelivery
-            DeliveryChannelRadio={DeliveryChannelRadio}
-            DeliveryRecipientsBox={DeliveryRecipientsBox}
             deliveryEmailSubject={this.state.deliveryEmailSubject}
             onChangeDeliveryEmailSubject={this.onChangeDeliveryEmailSubject}
             deliveryEmailBody={this.state.deliveryEmailBody}
@@ -489,7 +396,7 @@ export class CreateReport extends React.Component {
           <EuiFlexGroup justifyContent="flexEnd">
             <EuiFlexItem grow={false}>
               <EuiButtonEmpty
-                onClick={() => {window.location.assign('reporting#/')}}
+                onClick={() => {window.location.assign('opendistro-reporting#/')}}
               >
                 Cancel
               </EuiButtonEmpty>
